refactor as stand alone app
