+++ css/p3table.css
@@ -1,0 +1,45 @@
+/*ITC250 WN19 P3:Food Truck
+Group 4 (Esteban, Yonatan, Thom, Jeneva)
+
+Objective:  1.) CSS code for tabling Names, Descriptions, and Prices.
+            2.) Background Color
+            3.) Font
+
+*/
+
+.table, .table td, .table th {
+    border: 1px solid #809FE1;
+}
+
+.table {
+    border-collapse: collapse;
+    width: 100%;
+    text-align: center;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.table th {
+    height: 50px;
+    text-align: left;
+}
+
+.container {
+    background-color: antiquewhite;
+    border: 1px solid #809FE1;
+    width: 98%;
+    margin: 1%;
+    text-align: center
+}
+
+.table tr{
+    font-size: 1.750em;
+    font-family: "Times New Roman", Times, serif;
+    text-align: text-align: left;
+}
+
+.table td{
+    font-size: 0.875em;
+    font-family: "Times New Roman",
+        Times, serif;
+}
