+++ includes/footer.php
@@ -1,0 +1,11 @@
+ 
+<!-- START footer.php -->
+        <footer>
+            <BR />
+            <hr class="footer">
+            <p><small>&copy; 2018 - <?=date("Y")?> by Group, All Rights Reserved ~ <a href="https://validator.w3.org/check/referer" target="_blank">Valid HTML5</a> ~ <a href="https://jigsaw.w3.org/css-validator/check?uri=referer" target="_blank">Valid CSS</a></small></p>
+        </footer>
+<!-- END WRAPPER -->
+
+</body>
+</html>
