+++ includes/header.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * header_inc.php provides the initial HTML for our site files 
+ *
+ * @package nmCommon
+ * @author Bill Newman <williamnewman@gmail.com>
+ * @version 2.091 2011/06/17
+ * @link http://www.newmanix.com/ 
+ * @license https://www.apache.org/licenses/LICENSE-2.0
+ * @see meta_inc.php
+ * @see footer_inc.php 
+ * @todo none
+ */
+?>
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Food Truck</title>
+    <meta name="viewport" content="width=device-width" />
+    <meta name="robots" content="noindex,nofollow" />
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+	<link href="css/bootstrap.css" rel="stylesheet" type="text/css" />
+	<link href="css/bootstrap.min.css" rel="stylesheet" type="text/css" />
+	<link href="css/dbootstrap-theme.css" rel="stylesheet" type="text/css" />
+	<link href="css/bootstrap-theme.min.css" rel="stylesheet" type="text/css" />
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/js/bootstrap.min.js"></script>';
+<!--	<link href="css/p3table.css" rel="stylesheet" type="text/css" />-->
+	<style type="text/css">
+    </style>
+</head>
+
+		<!-- end of header include file -->
