+++ includes/items.php
@@ -1,0 +1,35 @@
+<?php
+//items.php just a list
+$myItem = new Item(1,"Taco","Our Tacos are awesome!",4.95);
+$myItem->addExtra("Sour Cream");
+$myItem->addExtra("Cheese");
+$myItem->addExtra("Guacamole");
+$config->items[] = $myItem;
+
+$myItem = new Item(2,"Sundae","Our Sundaes are awesome!",3.95);
+$myItem->addExtra("Sprinkles");
+$myItem->addExtra("Chocolate Sauce");
+$myItem->addExtra("Nuts");
+$config->items[] = $myItem;
+
+$myItem = new Item(3,"Salad","Our Salads are awesome!",5.95);
+$myItem->addExtra("Croutons");
+$myItem->addExtra("Bacon");
+$myItem->addExtra("Lemon Wedges");
+$myItem->addExtra("Avacado");
+$config->items[] = $myItem;
+
+$myItem = new Item(4,"Hot Tamale","Our Tamales are awesome!",7.95);
+$myItem->addExtra("Salsa");
+$myItem->addExtra("Guacamole");
+$myItem->addExtra("Habenaro Peppers");
+$myItem->addExtra("Sour Cream");
+$config->items[] = $myItem;
+
+$myItem = new Item(5,"Chile Relleno","Chile Relleno are awesome!",6.95);
+$myItem->addExtra("Salsa");
+$myItem->addExtra("Guacamole");
+$myItem->addExtra("Habenaro Peppers");
+$myItem->addExtra("Sour Cream");
+$config->items[] = $myItem;
+?>
